pekkanen segmenter
