+++ neurokit/bio/__init__.py
@@ -4,6 +4,7 @@ bio submodule.
 
 from .bio_data import *
 from .bio_rsp import *
+from .bio_ecg_preprocessing import *
 from .bio_ecg import *
 from .bio_eda import *
 from .bio_emg import *
